backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=24.4.2",
    "boto3>=1.42.59",
    "django>=6.0.2",
    "django-allauth[headless]>=65.14.3",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
    "django-filter>=25.2",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-nested-routers>=0.95.0",
    "drf-yasg>=1.21.15",
    "gunicorn>=25.1.0",
    "isort>=5.13.2",
    "pillow>=12.1.1",
    "pre-commit>=3.6.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "sentry-sdk[django]>=2.53.0",
    "whitenoise>=6.12.0",
]
